Cleanup commented code and add some comments
