+++ app/src/main/java/com/example/pharmscan/ui/Dialog/DelHostCompName.kt
@@ -11,7 +11,7 @@ import androidx.compose.ui.unit.dp
 fun DeleteHostComputerAlert(
         hostComp: String,
         showDialog: Boolean,
-        onDismiss: () -> Unit,
+        onDel: () -> Unit,
         onCancel: () -> Unit
     ) {
 
@@ -53,69 +53,10 @@ fun DeleteHostComputerAlert(
                             )
                         }
                         Button(
-                            onClick = { onDismiss() }
+                            onClick = { onDel() }
                         ) {
                             Text(
                                 text = " Delete ",
-                                style = MaterialTheme.typography.subtitle1
-                            )
-                        }
-                    }
-                }
-            }
-        )
-    }
-}
-
-@Composable
-fun CancelCollDataRecord(
-    collDataRec: String,
-    showDialog: Boolean,
-    onDismiss: () -> Unit
-) {
-
-    if (showDialog) {
-        AlertDialog(
-            onDismissRequest = {
-                onDismiss()
-            },
-            modifier = Modifier.size(200.dp, 200.dp),
-            title = {
-                Text(
-                    text = "Cancel?",
-                    style = MaterialTheme.typography.h6
-                )
-            },
-            text = {
-                Text(
-                    text = collDataRec,
-                    style = MaterialTheme.typography.h5
-                )
-            },
-            buttons = {
-                Column(
-                    horizontalAlignment = Alignment.End,
-                    modifier = Modifier
-                        .padding(all = 8.dp)
-                        .fillMaxWidth()
-                ) {
-                    Spacer(modifier = Modifier.height(35.dp))
-                    Row(
-                        verticalAlignment = Alignment.Bottom
-                    ) {
-                        TextButton(
-                            onClick = { onDismiss() }
-                        ) {
-                            Text(
-                                text = "Ignore",
-                                style = MaterialTheme.typography.subtitle1
-                            )
-                        }
-                        Button(
-                            onClick = { onDismiss() }
-                        ) {
-                            Text(
-                                text = " Cancel ",
                                 style = MaterialTheme.typography.subtitle1
                             )
                         }
