+++ app/version.properties
@@ -1,3 +1,3 @@
-#Thu Sep 02 16:04:36 EDT 2021
-VERSION_BUILD=682
+#Fri Sep 03 15:00:33 EDT 2021
+VERSION_BUILD=688
 VERSION_NUMBER=1.1
